fulgurant::auth

Module handlers

Structs

ForcePasswordUpdateRequest

ForgotPasswordStep1Request

ForgotPasswordStep2Request

ForgotPasswordStep3Request

LoginRequest

RegisterStep1Request

RegisterStep2Request

ResendCodeQuery

Functions

can_register

Checks if the registration is allowed

force_password_update

POST /force-password-update - Updates the password and clears the force_password_update flag

forgot_password_step_1

POST /auth/forgot-password - Step 1: Send verification code to email and shows the verification code form

forgot_password_step_2

POST /auth/forgot-password/verify - Step 2: Verify code and show the password form

forgot_password_step_3

POST /auth/forgot-password/reset - Step 3: Reset the password and show the success message

get_force_password_update_page

GET /force-password-update - Returns the force password update page

get_forgot_password_page

GET /auth/forgot-password - Returns the forgot password page

get_login_page

GET /login - Returns the login page

get_register_page

GET /register - Returns the register page

hash_password

Hashes a password

login

POST /login - Logs in a user

logout

POST /logout - Logs out a user

register_step_1

POST /register/step-1 - Registers a user, sends a verification code to the email address and shows the form for the second step of the registration process

register_step_2

POST /register/step-2 - Verifies the verification code and marks the user as verified

resend_forgot_password_code

GET /auth/forgot-password/resend - Resend the verification code for password reset